% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importData.R
\name{danStatGET}
\alias{danStatGET}
\title{A wrapper around httr::GET, providing the base url for Danmarks Statistik and
and sends a GET request based on the given inputs.}
\usage{
danStatGET(call, subject = NULL, table = NULL, variables = NULL)
}
\arguments{
\item{call}{A string with inputs. One of c("subjects", "tables", "tableinfo",
"data") for retrieving subjects, tables, table meta data and data
respectively.}

\item{variables}{A list containing various variables of interest}

\item{subjects}{A vector of strings providing specific subjects of interest.}

\item{tables}{A vector of strings providing specific tables.}
}
\value{
A data frame of the desired data.
}
\description{
A wrapper around httr::GET, providing the base url for Danmarks Statistik and
and sends a GET request based on the given inputs.
}

