% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_metadata.R
\name{get_table_metadata}
\alias{get_table_metadata}
\title{Title}
\usage{
get_table_metadata(table_id, variables_only = FALSE, language = c("en", "da"))
}
\arguments{
\item{table_id}{Table identifier, e.g. "folk1a"}

\item{variables_only}{If \code{TRUE} returns only information about the variables in the table}

\item{language}{Language for the return object. Default = \code{"en"}}
}
\value{
A list with information about the table, like documentation url, variable description, etc. If \code{variables_only = TRUE}, returns a data frame with variable information.
}
\description{
Title
}
\examples{
# Get table metadata for a given table
table_meta <- get_table_metadata(table_id = "folk1c") # a list

# Get only information about the variables in the table
table_meta_vars <- get_table_metadata(table_id = "folk1c", variables_only = TRUE) # a data frame
}
