% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Function importing data from the API of Danmarks Statistik. A wrapper around
httr.}
\usage{
importData(monthFrom, monthTo)
}
\arguments{
\item{monthFrom}{A date string in the format of (%Y-%m-%d)}

\item{monthTo}{A date string in the format of (%Y-%m-%d)}
}
\value{
A data frame of the requested data.
}
\description{
Function importing data from the API of Danmarks Statistik. A wrapper around
httr.
}

