% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importData.R
\name{importSubjects}
\alias{importSubjects}
\title{Function importing the various subjects from Danmarks Statistik.}
\usage{
importSubjects()
}
\value{
A vector containing the subjects
}
\description{
Function importing the various subjects from Danmarks Statistik.
}

