% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importData.R
\name{importTables}
\alias{importTables}
\title{Function importing the various tables from Danmarks statistik.}
\usage{
importTables(subject)
}
\arguments{
\item{subject}{A string with a subject from the subjects list.}
}
\value{
A vector of strings containing the table names.
}
\description{
Function importing the various tables from Danmarks statistik.
}

