% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotData}
\alias{plotData}
\title{Function providing plots for the various data available from Danmarks
statistik}
\usage{
plotData(inputFrame, unit, commodityGroup)
}
\value{
A plot element
}
\description{
Function providing plots for the various data available from Danmarks
statistik
}

