% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data for a particular table and variable selection}
\usage{
get_data(table_id, variables, language = c("en", "da"))
}
\arguments{
\item{table_id}{Table identifier, e.g. "folk1a"}

\item{variables}{A list with variable code-values pairs. Each code-values pair should be a named list with names "code" and "values".}

\item{language}{Language for the return object. Default = \code{"en"}}
}
\value{
A data frame
}
\description{
Get data for a particular table and variable selection
}
